+++ esp_h264/sw/include/esp_h264_dec_sw.h
@@ -26,9 +26,9 @@ typedef esp_h264_dec_cfg_t esp_h264_dec_cfg_sw_t;
 
 /**
  * @brief  This function is used to create a new instance of the `esp_h264_dec_t` data structure,
- *         which represents a single-streams H.264 decoder in software. The decoder is from tinyh264
+ *         which represents a single-stream H.264 decoder in software. The decoder is from tinyh264
  *
- * @note  The group of picture(GOP) will be updated in intra frame
+ * @note  The group of pictures (GOP) will be updated in intra frame
  *
  * @param[in]   cfg      It is a pointer to the `esp_h264_dec_cfg_sw_t` structure, which contains the configuration settings for the decoder
  * @param[out]  out_dec  It is a double pointer to the `esp_h264_dec_t` structure, which will store the created decoder instance